+++ src/main.rs
@@ -1,3 +1,5 @@
+use std::fs::File;
+
 use clap::Parser;
 use color_eyre::{
     Result,
@@ -12,6 +14,10 @@ use mqttui::{
 use petname::petname;
 use ratatui::{Terminal, prelude::Backend};
 use rumqttc::{AsyncClient, EventLoop, MqttOptions};
+use tracing::info;
+use tracing_subscriber::{
+    Layer, filter::filter_fn, fmt::format::FmtSpan, layer::SubscriberExt, util::SubscriberInitExt,
+};
 use url::Url;
 
 /// Mqtt TUI
@@ -20,7 +26,7 @@ use url::Url;
 #[derive(Debug, Parser)]
 struct Args {
     /// The URL of the MQTT broker to connect to
-    #[arg(short('b'), long, default_value = "mqtt://localhost:1883", value_parser = validate_url)]
+    #[arg(default_value = "mqtt://localhost:1883", value_parser = validate_url)]
     broker: Url,
 
     /// Immediately quit, when the connection to the broker is lost
@@ -33,6 +39,10 @@ struct Args {
     /// Print the path of where the config file is read from
     #[arg(long)]
     print_config_path: bool,
+
+    /// Print the path of where the log file is placed
+    #[arg(long)]
+    print_log_path: bool,
 }
 
 #[tokio::main]
@@ -44,38 +54,59 @@ async fn main() -> Result<()> {
         return Ok(());
     }
 
-    let mut terminal = ratatui::init();
-    // use ratatui::backend::TestBackend;
-    // let mut terminal = Terminal::new(TestBackend::new(10, 10)).unwrap();
+    if args.print_log_path {
+        println!("{}", Config::log()?.display());
+        return Ok(());
+    }
 
+    let mut terminal = ratatui::init();
     let result = run(args.broker, &mut terminal, !args.quit).await;
+
     ratatui::restore();
     result?;
     Ok(())
 }
 
-async fn init(broker: &Url) -> (AsyncClient, EventLoop) {
+async fn init(broker: &Url) -> Result<(AsyncClient, EventLoop)> {
     let hook = std::panic::take_hook();
     std::panic::set_hook(Box::new(move |panic_info| {
         ratatui::restore();
         hook(panic_info);
     }));
 
+    tracing_subscriber::registry()
+        .with(
+            tracing_subscriber::fmt::layer()
+                .with_span_events(FmtSpan::NEW)
+                .compact()
+                .with_file(true)
+                .with_line_number(true)
+                .with_target(false)
+                .with_writer(File::create(Config::log()?)?)
+                // rumqttc is pretty verbose, ignore it
+                .with_filter(filter_fn(|meta| meta.target() != "rumqttc")),
+        )
+        .init();
+    info!("Started MqtTUI: {broker}");
+
     let name = env!("CARGO_PKG_NAME");
     let name = format!("{name}-{}", petname(2, "-").unwrap());
 
     let mut options = MqttOptions::new(
         name,
-        broker.host_str().unwrap(),
+        broker
+            .host_str()
+            .ok_or(eyre!("Failed to find out host name from broker URL"))
+            .wrap_err(broker.clone())?,
         broker.port_or_known_default().unwrap_or(1883),
     );
     options.set_max_packet_size(1000000, 1024);
     let (client, eventloop) = AsyncClient::new(options, 10);
-    (client, eventloop)
+    Ok((client, eventloop))
 }
 
 async fn run<B: Backend>(broker: Url, terminal: &mut Terminal<B>, reconnect: bool) -> Result<()> {
-    let (client, eventloop) = init(&broker).await;
+    let (client, eventloop) = init(&broker).await?;
     let mut events = events::start(client, eventloop).await?;
     let mut model = Model::new(broker.clone())?;
     while !model.shutdown {
@@ -83,7 +114,7 @@ async fn run<B: Backend>(broker: Url, terminal: &mut Terminal<B>, reconnect: boo
         let rendering = event.is_render();
 
         if !reconnect && event.is_disconnect() {
-            return Err(eyre!(broker)).context("Connection to broker lost");
+            return Err(eyre!(broker)).wrap_err("Connection to broker lost");
         }
 
         model.update(event);
@@ -96,5 +127,9 @@ async fn run<B: Backend>(broker: Url, terminal: &mut Terminal<B>, reconnect: boo
 }
 
 fn validate_url(arg: &str) -> Result<Url, url::ParseError> {
-    Url::parse(arg).or_else(|_| format!("{arg}:1883").parse())
+    let url = Url::parse(arg).or_else(|_| format!("{arg}:1883").parse())?;
+    if url.host().is_none() {
+        return Url::parse(&format!("mqtt://{arg}")).or_else(|_| format!("{arg}:1883").parse());
+    }
+    Ok(url)
 }