+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Make broker URL positional argument
 * Fix bug where color for retained messages is lost
 * Fix bug where `<host>:<port>` was not allowed
 * Add logging to `~/.cache/mqttui/mqttui.log`