+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fix bug where color for retained messages is lost
 * Fix bug where `<host>:<port>` was not allowed
 * Add logging to `~/.cache/mqttui/mqttui.log`
+* Add message history
 
 ## 1.1.0